写真と動画のレコーダーを分離。プレビュー(MJPEG)中に写真を撮影できるように修正。
