+++ dConnectDevicePlugin/dConnectDeviceHost/app/src/main/java/org/deviceconnect/android/deviceplugin/host/recorder/camera/Camera2MJPEGPreviewServer.java
@@ -6,40 +6,17 @@
  */
 package org.deviceconnect.android.deviceplugin.host.recorder.camera;
 
-import android.content.Context;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.graphics.Matrix;
-import android.graphics.Rect;
-import android.graphics.YuvImage;
-import android.hardware.Camera;
-import android.hardware.camera2.CameraDevice;
-import android.support.annotation.NonNull;
-
 import org.deviceconnect.android.deviceplugin.host.recorder.AbstractPreviewServerProvider;
 import org.deviceconnect.android.deviceplugin.host.recorder.util.MixedReplaceMediaServer;
-
-import java.io.ByteArrayOutputStream;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-
-import static org.deviceconnect.android.deviceplugin.host.recorder.camera.CameraOverlay.FACING_DIRECTION_BACK;
-import static org.deviceconnect.android.deviceplugin.host.recorder.camera.CameraOverlay.FACING_DIRECTION_FRONT;
 
 
 class Camera2MJPEGPreviewServer implements CameraPreviewServer {
 
     private static final String MIME_TYPE = "video/x-mjpeg";
 
-    private HostDeviceCamera2Recorder mCameraRecorder;
-
     private final AbstractPreviewServerProvider mServerProvider;
 
     private final Object mLockObj = new Object();
-
-    private final Context mContext;
-
-    private CameraDevice mCamera;
 
     private MixedReplaceMediaServer mServer;
 
@@ -58,10 +35,7 @@ class Camera2MJPEGPreviewServer implements CameraPreviewServer {
         void onStop();
     }
 
-    Camera2MJPEGPreviewServer(final Context context,
-                              final HostDeviceCamera2Recorder recorder) {
-        mContext = context;
-        mCameraRecorder = recorder;
+    Camera2MJPEGPreviewServer(final Camera2PhotoRecorder recorder) {
         mServerProvider = recorder;
     }
 
@@ -91,6 +65,10 @@ class Camera2MJPEGPreviewServer implements CameraPreviewServer {
                 mServer.setContentType("image/jpg");
                 mServer.setCallback(mMediaServerCallback);
                 uri = mServer.start();
+
+                if (mPreviewServerListener != null) {
+                    mPreviewServerListener.onStart();
+                }
             } else {
                 uri = mServer.getUrl();
             }
@@ -106,6 +84,10 @@ class Camera2MJPEGPreviewServer implements CameraPreviewServer {
                 mServer = null;
             }
             mServerProvider.hideNotification();
+
+            if (mPreviewServerListener != null) {
+                mPreviewServerListener.onStop();
+            }
         }
     }
 }
